Finished first full version of CLI. Improved some scripts to be more robust for more general use cases.

+++ Tensorflow/visualization.py
@@ -7,15 +7,9 @@ them to the output_folder. Alongside each image either a json file or an xml fil
 data has to be provided. 
 @author: johan
 """
+print("Loading libraries...")
 
 from utils import constants
-input_folder = constants.input_folders[0]
-input_folder = "C:/Users/johan/Desktop/output/images/annotated_ortho_photos"
-input_folder = "C:/Users/johan/Desktop/bothtablets"
-output_folder = constants.vis_im
-
-
-
 from object_detection.utils import visualization_utils
 from PIL import Image
 from utils import file_utils
@@ -24,51 +18,38 @@ import os
 import progressbar
 
 
-def get_color_for_index(index):
-    return STANDARD_COLORS[index]
-    #label = list(matplotlib.colors.cnames.keys())[index]
-    #return label
 
 
-def draw_bounding_boxes(input_folder, output_folder):
+def draw_bounding_boxes(input_folder, output_folder, with_name_info=True, clean_output_folder=True):
     flowers = []
     images = file_utils.get_all_images_in_folder(input_folder)
     print("Drawing bounding boxes on images:")
-    file_utils.delete_folder_contents(output_folder)
+    if clean_output_folder:
+        file_utils.delete_folder_contents(output_folder)
     for i in progressbar.progressbar(range(len(images))):
         image_path = images[i]
         image = Image.open(image_path)
         
-        annotation_path = image_path[:-4] + "_annotations.json"
-        annotation_path_xml = image_path[:-4] + ".xml"
-        if annotation_path and os.path.isfile(annotation_path):
-            annotation_data = file_utils.read_json_file(annotation_path)
-            if not annotation_data:
-                continue
-            annotations = annotation_data["annotatedFlowers"]
-            for flower in annotations:
-                flower_name = flower_info.clean_string(flower["name"])
-                [top,left,bottom,right] = flower_info.get_bbox(flower)
-                if not flower_name in flowers:
-                    flowers.append(flower_name)
-                col = get_color_for_index(flowers.index(flower_name))
+        annotations = file_utils.get_annotations(image_path)
+        for flower in annotations:
+            flower_name = flower_info.clean_string(flower["name"])
+            [top,left,bottom,right] = flower_info.get_bbox(flower)
+            if not flower_name in flowers:
+                flowers.append(flower_name)
+            col = get_color_for_index(flowers.index(flower_name))
+            if with_name_info:
                 visualization_utils.draw_bounding_box_on_image(image,top,left,bottom,right,display_str_list=[flower_name], color=col, use_normalized_coordinates=False, thickness=1)
-                
+            else:
+                visualization_utils.draw_bounding_box_on_image(image,top,left,bottom,right,display_str_list=[], color=col, use_normalized_coordinates=False, thickness=1)
         
-        elif annotation_path_xml and os.path.isfile(annotation_path_xml):
-            annotations = file_utils.get_annotations_from_xml(annotation_path_xml)
-            for flower in annotations:
-                flower_name = flower_info.clean_string(flower["name"])
-                [top,left,bottom,right] = flower["bounding_box"]
-                if not flower_name in flowers:
-                    flowers.append(flower_name)
-                col = get_color_for_index(flowers.index(flower_name))
-                visualization_utils.draw_bounding_box_on_image(image,top,left,bottom,right,display_str_list=[flower_name], color=col, use_normalized_coordinates=False, thickness=1)
-
         image_name = os.path.basename(image_path)
         image.save(os.path.join(output_folder,image_name))
     print("Done!")
     
+def get_color_for_index(index):
+    return STANDARD_COLORS[index]
+    #label = list(matplotlib.colors.cnames.keys())[index]
+    #return label
 
 STANDARD_COLORS = [
     'AliceBlue', 'Chartreuse', 'Aqua', 'Aquamarine', 'Azure', 'Beige', 'Bisque',
@@ -96,4 +77,10 @@ STANDARD_COLORS = [
     'WhiteSmoke', 'Yellow', 'YellowGreen'
 ]
 
-draw_bounding_boxes(input_folder,output_folder)
+if __name__ == '__main__':
+    
+    input_folder = constants.train_dir + "/images/train"
+    output_folder = constants.vis_im
+    
+    draw_bounding_boxes(input_folder,output_folder)
+
